+++ addons/docs/docs.js
@@ -1,0 +1,1 @@
+export {DocsPage, DocsContainer} from '@storybook/addon-docs/blocks';