+++ addons/knobs/configure.js
@@ -1,11 +1,1 @@
-import {withKnobs as addonWithKnobs} from '@storybook/addon-knobs';
-
-function withKnobs(opts) {
-	return addonWithKnobs({
-		timestamps: true, // Doesn't emit events while user is typing.
-		...opts
-	});
-}
-
-export default withKnobs;
-export {withKnobs};
+export {withKnobs} from '@storybook/addon-knobs';